Move glue.viewers.common.viz_client to glue.viewers.matplotlib.mpl_axes
